dev-grapqhql-schema-update
* Organize queries, mutations and types into different packages
