+++ repository/Favorite.js
@@ -30,6 +30,7 @@ class FavoritesRepository extends BaseRepository {
 
     try {
       const result = await dbQuery(queryConfig);
+
       return result.rows;
     } catch (error) {
       return error;
