+++ repository/Reply.js
@@ -35,6 +35,7 @@ class ReplyRepository extends BaseRepository {
 
     try {
       const result = await dbQuery(queryConfig);
+
       return result.rows;
     } catch (error) {
       stackLogger(error);
