+++ repository/Review.js
@@ -39,6 +39,7 @@ class ReviewRepository extends BaseRepository {
 
     try {
       const result = await dbQuery(queryConfig);
+
       return result.rows;
     } catch (error) {
       return error;
