+++ repository/User.js
@@ -32,6 +32,7 @@ class UserRepository extends BaseRepository {
 
     try {
       const result = await dbQuery(queryConfig);
+
       return result.rows;
     } catch (error) {
       return error;
